pyproject.toml
[project]
name = "oklab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "color-science>=0.4.5",
    "colorspacious>=1.1.2",
    "colour-science>=0.4.6",
    "matplotlib>=3.9.2",
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
]
